refactored package structure, refactored ExerciseSummaryValuesDecorator
