--- st-exerciseviewer/src/main/java/de/saring/exerciseviewer/parser/impl/garminfit/ExerciseSummaryCalculator.java
@@ -1,8 +1,0 @@
-package de.saring.exerciseviewer.parser.impl.garminfit;
-
-import de.saring.exerciseviewer.data.ExerciseSample;
-
-public interface ExerciseSummaryCalculator {
-	void visitSample(ExerciseSample sample);
-	void finished();
-}
