+++ st-exerciseviewer/src/main/java/de/saring/exerciseviewer/parser/impl/garminfit/summaryCalculators/ExerciseSummaryAggregator.java
@@ -1,0 +1,8 @@
+package de.saring.exerciseviewer.parser.impl.garminfit.summaryCalculators;
+
+import de.saring.exerciseviewer.data.ExerciseSample;
+
+public interface ExerciseSummaryAggregator {
+	void collect(ExerciseSample sample);
+	void aggregateAndUpdateExercise();
+}
