+++ st-exerciseviewer/src/main/java/de/saring/exerciseviewer/parser/impl/garminfit/summaryCalculators/MaxSpeedCalculator.java
@@ -1,9 +1,9 @@
-package de.saring.exerciseviewer.parser.impl.garminfit;
+package de.saring.exerciseviewer.parser.impl.garminfit.summaryCalculators;
 
 import de.saring.exerciseviewer.data.EVExercise;
 import de.saring.exerciseviewer.data.ExerciseSample;
 
-public class MaxSpeedCalculator implements ExerciseSummaryCalculator {
+public class MaxSpeedCalculator implements ExerciseSummaryAggregator {
 
 	private EVExercise exercise;
 	private float max = 0;
@@ -13,7 +13,7 @@ public class MaxSpeedCalculator implements ExerciseSummaryCalculator {
 	}
 
 	@Override
-	public void visitSample(ExerciseSample sample) {
+	public void collect(ExerciseSample sample) {
 		if (sample.getSpeed() > max) {
 			max = sample.getSpeed();
 		}
@@ -21,7 +21,7 @@ public class MaxSpeedCalculator implements ExerciseSummaryCalculator {
 	}
 
 	@Override
-	public void finished() {
+	public void aggregateAndUpdateExercise() {
 		if (exercise.getSpeed() != null) {
 			exercise.getSpeed().setSpeedMax(max);
 		}
